+++ backend/src/main/java/com/animo/animorise/repository/AnimalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface AnimalRepository extends JpaRepository<Animal, Long> {
-    List<Animal> findByOwnerId(Integer ownerId);
+    List<Animal> findByOwnerId(Long ownerId);
 }
